pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "MVP_INS"
version = "1.0.0"
description = "An insurance claim processing agent using LiveKit and OpenAI plugins."
requires-python = ">=3.10, <3.14"

dependencies = [
   "asyncio>=4.0.0",
   "asyncpg>=0.31.0",
   "livekit-agents[elevenlabs,silero,turn-detector]~=1.3",
   "livekit-plugins-noise-cancellation~=0.2",
   "livekit-plugins-openai>=1.3.12",
   "pip>=25.3",
   "pydantic>=2.12.5",
   "python-dotenv",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
